python-module/ed2k/pyproject.toml
[tool.poetry]
name = "ed2k"
version = "0.0.2.1"
description = "Python ed2k hash module."
authors = ["ChenyangGao <wosiwujm@gmail.com>"]
license = "MIT"
readme = "readme.md"
homepage = "https://github.com/ChenyangGao/web-mount-packs/tree/main/python-module/ed2k"
repository = "https://github.com/ChenyangGao/web-mount-packs/tree/main/python-module/ed2k"
keywords = ["ed2k", "hash"]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Development Status :: 5 - Production/Stable",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Software Development",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
include = [
    "LICENSE",
]

[tool.poetry.dependencies]
python = "^3.10"
iterdir = ">=0.0.6"
pycryptodome = "*"
python-filewrap = ">=0.2.6"

[tool.poetry.scripts]
python-ed2k = "ed2k.__main__:main"
ed2k = "ed2k.__main__:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.packages]]
include = "ed2k"
